Added authentication support for the HtmlUnitDriver using capabilities
Previously authentication could not be done with the
RemoteWebDriver, because it was only possible by overriding
the modifyWebDriver method of the HtmlUnitDriver. This commit
adds a WebClientCredentials object similar to the Proxy object
to configure credentials, which can be included in the
DesiredCapabilities just like the Proxy.

.NET bindings are included. Unit tests are not written yet and
should be added soon.